modified colorId shader

+++ libs/core/scenegraph/polygonaldrawable.cpp
@@ -71,7 +71,7 @@ void PolygonalDrawable::draw(
     program.use();
     program.setUniform(TRANSFORM_UNIFORM, transform);
 
-    glm::uvec4 id_vec = colorVectorFromId(m_id);
+    glm::vec4 id_vec = colorVectorFromId(m_id);
     program.setUniform(ID_UNIFORM, id_vec);
 
     glBindVertexArray( m_geometry->vao() );
@@ -93,9 +93,9 @@ void PolygonalDrawable::draw(
     program.release();
 }
 
-glm::uvec4 PolygonalDrawable::colorVectorFromId(const unsigned int & id)
+glm::vec4 PolygonalDrawable::colorVectorFromId(const unsigned int & id)
 {
-    glm::uvec4 id_vec;
+    glm::vec4 id_vec;
     id_vec.r = id % 256;
     id_vec.g = (id >> 8) % 256;
     id_vec.b = (id >> 16) % 256;
